+++ internal/service/ref/refs_test.go
@@ -6,7 +6,9 @@ import (
 	"io/ioutil"
 	"testing"
 
+	"github.com/golang/protobuf/ptypes/timestamp"
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -449,5 +451,114 @@ func TestEmptyFindLocalBranchesRequest(t *testing.T) {
 
 	if grpc.Code(recvError) != codes.InvalidArgument {
 		t.Fatal(recvError)
+	}
+}
+
+func deleteRemoteBranch(t *testing.T, repoPath, remoteName, branchName string) {
+	testhelper.MustRunCommand(t, nil, "git", "-C", repoPath, "update-ref", "-d",
+		"refs/remotes/"+remoteName+"/"+branchName)
+}
+
+func createRemoteBranch(t *testing.T, repoPath, remoteName, branchName, ref string) {
+	testhelper.MustRunCommand(t, nil, "git", "-C", repoPath, "update-ref",
+		"refs/remotes/"+remoteName+"/"+branchName, ref)
+}
+
+func TestSuccessfulFindAllBranchesRequest(t *testing.T) {
+	server := runRefServiceServer(t)
+	defer server.Stop()
+
+	remoteBranch := &pb.FindAllBranchesResponse_Branch{
+		Name: []byte("refs/remotes/origin/fake-remote-branch"),
+		Target: &pb.GitCommit{
+			Id:      "913c66a37b4a45b9769037c55c2d238bd0942d2e",
+			Subject: []byte("Files, encoding and much more"),
+			Author: &pb.CommitAuthor{
+				Name:  []byte("Dmitriy Zaporozhets"),
+				Email: []byte("<dmitriy.zaporozhets@gmail.com>"),
+				Date:  &timestamp.Timestamp{Seconds: 1393488896},
+			},
+			Committer: &pb.CommitAuthor{
+				Name:  []byte("Dmitriy Zaporozhets"),
+				Email: []byte("<dmitriy.zaporozhets@gmail.com>"),
+				Date:  &timestamp.Timestamp{Seconds: 1393488896},
+			},
+		},
+	}
+
+	createRemoteBranch(t, testRepoPath, "origin", "fake-remote-branch",
+		remoteBranch.Target.Id)
+	defer deleteRemoteBranch(t, testRepoPath, "origin", "fake-remote-branch")
+
+	request := &pb.FindAllBranchesRequest{Repository: testRepo}
+	client := newRefServiceClient(t)
+	c, err := client.FindAllBranches(context.Background(), request)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var branches []*pb.FindAllBranchesResponse_Branch
+	for {
+		r, err := c.Recv()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			t.Fatal(err)
+		}
+		branches = append(branches, r.GetBranches()...)
+	}
+
+	// It contains local branches
+	for name, target := range localBranches {
+		branch := &pb.FindAllBranchesResponse_Branch{
+			Name:   []byte(name),
+			Target: target,
+		}
+		assertContainsBranch(t, branches, branch)
+	}
+
+	// It contains our fake remote branch
+	assertContainsBranch(t, branches, remoteBranch)
+}
+
+func TestInvalidFindAllBranchesRequest(t *testing.T) {
+	server := runRefServiceServer(t)
+	defer server.Stop()
+
+	client := newRefServiceClient(t)
+	testCases := []struct {
+		description string
+		request     pb.FindAllBranchesRequest
+	}{
+		{
+			description: "Empty request",
+			request:     pb.FindAllBranchesRequest{},
+		},
+		{
+			description: "Invalid repo",
+			request: pb.FindAllBranchesRequest{
+				Repository: &pb.Repository{
+					StorageName:  "fake",
+					RelativePath: "repo",
+				},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Logf("test case: %v", tc.description)
+
+		c, err := client.FindAllBranches(context.Background(), &tc.request)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		var recvError error
+		for recvError == nil {
+			_, recvError = c.Recv()
+		}
+
+		testhelper.AssertGrpcError(t, recvError, codes.InvalidArgument, "")
 	}
 }